Fix login loop when session expires while trying to access a protected resource
