+++ xently/libraries/data-network/src/main/java/co/ke/xently/libraries/data/network/AccessTokenManager.kt
@@ -2,7 +2,8 @@ package co.ke.xently.libraries.data.network
 
 import io.ktor.client.HttpClient
 
-interface AccessTokenProvider {
+interface AccessTokenManager {
+    suspend fun clearUserSession()
     suspend fun getAccessToken(): String?
     suspend fun getFreshAccessToken(httpClient: HttpClient): String?
 }
