+++ xently/libraries/data-network/src/main/java/co/ke/xently/libraries/data/network/di/NetworkModule.kt
@@ -1,7 +1,7 @@
 package co.ke.xently.libraries.data.network.di
 
 import android.content.Context
-import co.ke.xently.libraries.data.network.AccessTokenProvider
+import co.ke.xently.libraries.data.network.AccessTokenManager
 import co.ke.xently.libraries.data.network.BaseURL
 import co.ke.xently.libraries.data.network.HttpClientFactory
 import dagger.Module
@@ -32,13 +32,13 @@ object NetworkModule {
         context: Context,
         json: Json,
         baseURL: BaseURL,
-        accessTokenProvider: AccessTokenProvider,
+        accessTokenManager: AccessTokenManager,
     ): HttpClient {
         return HttpClientFactory(
             context = context,
             json = json,
             baseURL = baseURL,
-            accessTokenProvider = accessTokenProvider,
+            accessTokenManager = accessTokenManager,
         )
     }
 }
